refactor: locate within physics related namespace
